Simplify ieq GQL filter
